Skip Forward and Backward
The controls now work. Used currentIndex to first find the current Song ID. From there we have an if statement either +1 to array of songs or -1 from array of songs, stored in currentIndex. We also have a fail safe with the (%) remainder operator, to ensure we can traverse the array both forwards and backwards without issues. We also have like a failstop set if we try to go backwards on the array[0], and if statement check that, and if it would resolve to -1, we rather update it to the array.length - 1.